Move truncate method to text module
